feat: move get_parallel_build_count and tests to util
