+++ craft_application/services/lifecycle.py
@@ -200,7 +200,7 @@ class LifecycleService(base.ProjectService):
                 cache_dir=self._cache_dir,
                 work_dir=self._work_dir,
                 ignore_local_sources=self._app.source_ignore_patterns,
-                parallel_build_count=self._get_parallel_build_count(),
+                parallel_build_count=util.get_parallel_build_count(self._app.name),
                 project_vars_part_name=self._project.adopt_info,
                 project_vars=self._project_vars,
                 track_stage_packages=True,
@@ -323,92 +323,6 @@ class LifecycleService(base.ProjectService):
             f"{self.__class__.__name__}({self._app!r}, {self._project!r}, "
             f"{work_dir=}, {cache_dir=}, {plan=}, **{self._manager_kwargs!r})"
         )
-
-    def _verify_parallel_build_count(
-        self, env_name: str, parallel_build_count: int | str
-    ) -> int:
-        """Verify the parallel build count is valid.
-
-        :param env_name: The name of the environment variable being checked.
-        :param parallel_build_count: The value of the variable.
-        :return: The parallel build count as an integer.
-        """
-        try:
-            parallel_build_count = int(parallel_build_count)
-        except ValueError as err:
-            raise errors.InvalidParameterError(
-                env_name, str(os.environ[env_name])
-            ) from err
-
-        # Ensure the value is valid positive integer
-        if parallel_build_count < 1:
-            raise errors.InvalidParameterError(env_name, str(parallel_build_count))
-
-        return parallel_build_count
-
-    def _get_parallel_build_count(self) -> int:
-        """Get the number of parallel builds to run.
-
-        The parallel build count is determined by the first available of the
-        following environment variables in the order:
-
-        - <APP_NAME>_PARALLEL_BUILD_COUNT
-        - CRAFT_PARALLEL_BUILD_COUNT
-        - <APP_NAME>_MAX_PARALLEL_BUILD_COUNT
-        - CRAFT_MAX_PARALLEL_BUILD_COUNT
-
-        where the MAX_PARALLEL_BUILD_COUNT variables are dynamically compared to
-        the number of CPUs, and the smaller of the two is used.
-
-        If no environment variable is set, the CPU count is used.
-        If the CPU count is not available for some reason, 1 is used as a fallback.
-        """
-        parallel_build_count = None
-
-        # fixed parallel build count environment variable
-        for env_name in [
-            (self._app.name + "_PARALLEL_BUILD_COUNT").upper(),
-            "CRAFT_PARALLEL_BUILD_COUNT",
-        ]:
-            if os.environ.get(env_name):
-                parallel_build_count = self._verify_parallel_build_count(
-                    env_name, os.environ[env_name]
-                )
-                emit.debug(
-                    f"Using parallel build count of {parallel_build_count} "
-                    f"from environment variable {env_name!r}"
-                )
-                break
-
-        # CPU count related max parallel build count environment variable
-        if parallel_build_count is None:
-            cpu_count = os.cpu_count() or 1
-            for env_name in [
-                (self._app.name + "_MAX_PARALLEL_BUILD_COUNT").upper(),
-                "CRAFT_MAX_PARALLEL_BUILD_COUNT",
-            ]:
-                if os.environ.get(env_name):
-                    parallel_build_count = min(
-                        cpu_count,
-                        self._verify_parallel_build_count(
-                            env_name, os.environ[env_name]
-                        ),
-                    )
-                    emit.debug(
-                        f"Using parallel build count of {parallel_build_count} "
-                        f"from environment variable {env_name!r}"
-                    )
-                    break
-
-            # Default to CPU count if no max environment variable is set
-            if parallel_build_count is None:
-                parallel_build_count = cpu_count
-                emit.debug(
-                    f"Using parallel build count of {parallel_build_count} "
-                    "from CPU count"
-                )
-
-        return parallel_build_count
 
     def _get_local_keys_path(self) -> Path | None:
         """Return a directory with public keys for package-repositories.
