+++ craft_application/util/__init__.py
@@ -32,6 +32,7 @@ from craft_application.util.snap_config import (
     is_running_from_snap,
 )
 from craft_application.util.string import humanize_list, strtobool
+from craft_application.util.system import get_parallel_build_count
 from craft_application.util.yaml import dump_yaml, safe_yaml_load
 
 __all__ = [
@@ -52,4 +53,5 @@ __all__ = [
     "dump_yaml",
     "safe_yaml_load",
     "retry",
+    "get_parallel_build_count",
 ]
