+++ pyproject.toml
@@ -81,7 +81,7 @@ docs = [
     "sphinx-lint==0.9.1",
 ]
 apt = [
-    "python-apt>=2.4.0;sys_platform=='linux'"
+    "python-apt @ https://launchpad.net/ubuntu/+archive/primary/+sourcefiles/python-apt/2.4.0ubuntu2/python-apt_2.4.0ubuntu2.tar.xz ; sys_platform == 'linux'"
 ]
 
 [build-system]
