+++ tests/unit/services/test_lifecycle.py
@@ -614,160 +614,6 @@ def test_lifecycle_package_repositories(
 
 # endregion
 
-# region parallel build count tests
-
-
-@pytest.mark.parametrize(
-    ("env_dict", "cpu_count", "expected"),
-    [
-        (
-            {},
-            None,
-            1,
-        ),
-        (
-            {},
-            100,
-            100,
-        ),
-        (
-            {"TESTCRAFT_PARALLEL_BUILD_COUNT": "100"},
-            1,
-            100,
-        ),
-        (
-            {"CRAFT_PARALLEL_BUILD_COUNT": "200"},
-            1,
-            200,
-        ),
-        (
-            {
-                "TESTCRAFT_MAX_PARALLEL_BUILD_COUNT": "100",
-            },
-            50,
-            50,
-        ),
-        (
-            {
-                "CRAFT_MAX_PARALLEL_BUILD_COUNT": "100",
-            },
-            80,
-            80,
-        ),
-        (
-            {
-                "TESTCRAFT_PARALLEL_BUILD_COUNT": "100",
-                "CRAFT_PARALLEL_BUILD_COUNT": "200",
-            },
-            1,
-            100,
-        ),
-        (
-            {
-                "TESTCRAFT_MAX_PARALLEL_BUILD_COUNT": "100",
-                "CRAFT_MAX_PARALLEL_BUILD_COUNT": "200",
-            },
-            150,
-            100,
-        ),
-        (
-            {
-                "TESTCRAFT_MAX_PARALLEL_BUILD_COUNT": "100",
-                "CRAFT_MAX_PARALLEL_BUILD_COUNT": "200",
-            },
-            None,
-            1,
-        ),
-        (
-            {
-                "TESTCRAFT_PARALLEL_BUILD_COUNT": "100",
-                "CRAFT_PARALLEL_BUILD_COUNT": "200",
-                "TESTCRAFT_MAX_PARALLEL_BUILD_COUNT": "300",
-                "CRAFT_MAX_PARALLEL_BUILD_COUNT": "400",
-            },
-            150,
-            100,
-        ),
-    ],
-)
-def test_get_parallel_build_count(
-    monkeypatch, mocker, fake_parts_lifecycle, env_dict, cpu_count, expected
-):
-    mocker.patch("os.cpu_count", return_value=cpu_count)
-    for env_dict_key, env_dict_value in env_dict.items():
-        monkeypatch.setenv(env_dict_key, env_dict_value)
-
-    assert fake_parts_lifecycle._get_parallel_build_count() == expected
-
-
-@pytest.mark.parametrize(
-    ("env_dict", "cpu_count"),
-    [
-        (
-            {
-                "TESTCRAFT_PARALLEL_BUILD_COUNT": "abc",
-            },
-            1,
-        ),
-        (
-            {
-                "CRAFT_PARALLEL_BUILD_COUNT": "-",
-            },
-            1,
-        ),
-        (
-            {
-                "TESTCRAFT_MAX_PARALLEL_BUILD_COUNT": "*",
-            },
-            1,
-        ),
-        (
-            {
-                "CRAFT_MAX_PARALLEL_BUILD_COUNT": "$COUNT",
-            },
-            1,
-        ),
-        (
-            {
-                "TESTCRAFT_PARALLEL_BUILD_COUNT": "0",
-            },
-            1,
-        ),
-        (
-            {
-                "CRAFT_PARALLEL_BUILD_COUNT": "-1",
-            },
-            1,
-        ),
-        (
-            {
-                "TESTCRAFT_MAX_PARALLEL_BUILD_COUNT": "5.6",
-            },
-            1,
-        ),
-        (
-            {
-                "CRAFT_MAX_PARALLEL_BUILD_COUNT": "inf",
-            },
-            1,
-        ),
-    ],
-)
-def test_get_parallel_build_count_error(
-    monkeypatch, mocker, fake_parts_lifecycle, env_dict, cpu_count
-):
-    mocker.patch("os.cpu_count", return_value=cpu_count)
-    for env_dict_key, env_dict_value in env_dict.items():
-        monkeypatch.setenv(env_dict_key, env_dict_value)
-
-    with pytest.raises(
-        InvalidParameterError, match=r"^Value '.*' is invalid for parameter '.*'.$"
-    ):
-        fake_parts_lifecycle._get_parallel_build_count()
-
-
-# endregion
-
 # region project variables
